Make DetailedContent protocol iOS-only.
